Rename patch file
